backend/requirements.txt
Flask==3.0.0
Mastodon.py==1.8.1
azure-functions